requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2-binary
motor
python-dotenv
openai
google-genai
mistralai
scikit-learn
pandas
numpy<2.0.0
httpx
python-multipart
fpdf2
